# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssuranceDecision": {
        "enum": [
          "execute",
          "execute_flagged",
          "needs_human"
        ],
        "title": "AssuranceDecision",
        "type": "string"
      },
      "AssuranceEvaluationOut": {
        "additionalProperties": false,
        "properties": {
          "action_type": {
            "title": "Action Type",
            "type": "string"
          },
          "blocking": {
            "items": {
              "$ref": "#/components/schemas/CheckName"
            },
            "title": "Blocking",
            "type": "array"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/CheckResultOut"
            },
            "title": "Checks",
            "type": "array"
          },
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/AssuranceDecision"
          },
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "human_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanDecisionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "plan_task_id": {
            "title": "Plan Task Id",
            "type": "integer"
          },
          "risk_tier": {
            "$ref": "#/components/schemas/RiskTier"
          },
          "warn_permitted_by_config": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warn Permitted By Config"
          }
        },
        "required": [
          "id",
          "plan_task_id",
          "action_type",
          "decision",
          "risk_tier",
          "evaluated_at",
          "config_version",
          "config_hash"
        ],
        "title": "AssuranceEvaluationOut",
        "type": "object"
      },
      "AssuranceResponse": {
        "additionalProperties": false,
        "properties": {
          "awaiting_approval_count": {
            "default": 0,
            "title": "Awaiting Approval Count",
            "type": "integer"
          },
          "config_hash": {
            "title": "Config Hash",
            "type": "string"
          },
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/AssuranceEvaluationOut"
            },
            "title": "Evaluations",
            "type": "array"
          },
          "incident_reference": {
            "title": "Incident Reference",
            "type": "string"
          }
        },
        "required": [
          "incident_reference",
          "config_version",
          "config_hash"
        ],
        "title": "AssuranceResponse",
        "type": "object"
      },
      "CheckName": {
        "enum": [
          "evidence_complete",
          "sources_fresh",
          "entities_valid",
          "policy_compliant",
          "no_conflicts",
          "action_risk"
        ],
        "title": "CheckName",
        "type": "string"
      },
      "CheckResultOut": {
        "additionalProperties": false,
        "properties": {
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "name": {
            "$ref": "#/components/schemas/CheckName"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "reason_code": {
            "$ref": "#/components/schemas/ReasonCode"
          },
          "state": {
            "$ref": "#/components/schemas/CheckState"
          },
          "tier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskTier"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "name",
          "state",
          "reason_code"
        ],
        "title": "CheckResultOut",
        "type": "object"
      },
      "CheckState": {
        "description": "Three states, deliberately. A WARN must never collapse into a boolean.",
        "enum": [
          "PASS",
          "WARN",
          "FAIL"
        ],
        "title": "CheckState",
        "type": "string"
      },
      "DecisionRequest": {
        "additionalProperties": false,
        "description": "An operator approving or rejecting one evaluation.",
        "properties": {
          "actor_id": {
            "default": "operator-1",
            "maxLength": 64,
            "title": "Actor Id",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/HumanDecisionType"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
      "DecisionResponse": {
        "additionalProperties": false,
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "assurance_id": {
            "title": "Assurance Id",
            "type": "integer"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/HumanDecisionType"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "replayed": {
            "default": false,
            "title": "Replayed",
            "type": "boolean"
          }
        },
        "required": [
          "assurance_id",
          "decision",
          "actor_id",
          "reason",
          "decided_at"
        ],
        "title": "DecisionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HumanDecisionOut": {
        "additionalProperties": false,
        "description": "An immutable operator response, keyed to exactly one evaluation.",
        "properties": {
          "actor_id": {
            "title": "Actor Id",
            "type": "string"
          },
          "decided_at": {
            "format": "date-time",
            "title": "Decided At",
            "type": "string"
          },
          "decision": {
            "$ref": "#/components/schemas/HumanDecisionType"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "id",
          "decision",
          "actor_id",
          "reason",
          "decided_at"
        ],
        "title": "HumanDecisionOut",
        "type": "object"
      },
      "HumanDecisionType": {
        "enum": [
          "approved",
          "rejected"
        ],
        "title": "HumanDecisionType",
        "type": "string"
      },
...
      "ReasonCode": {
        "description": "Stable codes. The UI maps these to copy; it never parses free text.",
        "enum": [
          "OK",
          "MISSING_EVIDENCE",
          "SOURCE_STALE",
          "SOURCE_MISSING_TIMESTAMP",
          "ENTITY_NOT_FOUND",
          "ENTITY_STATE_MISMATCH",
          "POLICY_CONSTRAINT_BREACH",
          "POLICY_PACK_UNAVAILABLE",
          "DUPLICATE_ACTION",
          "CAPACITY_UNAVAILABLE",
          "HUMAN_APPROVAL_REQUIRED",
          "UNKNOWN_ACTION_TYPE",
          "UNKNOWN_RULE_OPERATOR",
          "CONFIG_MISSING",
          "MISSING_REQUIRED_FACT"
        ],
        "title": "ReasonCode",
        "type": "string"
      },
...
      "RiskTier": {
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskTier",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Autonomous operating layer for airline disruption recovery",
    "title": "TravelOps AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assurance/{assurance_id}/decision": {
      "post": {
        "description": "Record an immutable operator response for one evaluation.\n\nIdempotency here comes from the schema rather than a cache: `human_decision.assurance_id`\nis unique, so one evaluation can only ever carry one decision. Re-posting the same\ndecision returns the original; posting a different one is a conflict, because an\noperator response is a record of what someone decided and not a mutable setting.",
        "operationId": "record_decision_api_v1_assurance__assurance_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "assurance_id",
            "required": true,
            "schema": {
              "title": "Assurance Id",
              "type": "integer"
            }
          },
          {
            "description": "A replay returns the original decision.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A replay returns the original decision.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operator approve or reject, with a reason",
        "tags": [
          "assurance"
        ]
      }
    },
...
    "/api/v1/incidents/{incident_id}/assurance": {
      "get": {
        "operationId": "get_assurance_api_v1_incidents__incident_id__assurance_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssuranceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gate evaluations and their failed checks",
        "tags": [
          "assurance"
        ]
      }
    },
...
  }
}